C++ and MatLab files
I used MatLab to initially understand the algorithm and then completed
with C++. The C++ file is fully templated and requires armadillo
(http://arma.sourceforge.net/) to work. Armadillo was chosen as the
math library due to it's similar syntax to MatLab, which the end users
are more familiar with.